Fix for Office 2013 and output.txt path
Add a similar way to check office 2013 as the one for office 2016, there is no registry key, we must use the OSPP.vbs script to get the key.

Fix pah for reading the output.txt as the current one is inconsistent with the way we create it (%USERPROFILE% vs C:\users\%USERNAME%). In a domain if you have the same name for an account in the domain and out of it, this will leads to a wrong path (domain won't be appended in the path)